requirements.txt
pytest
allure-pytest
selene>=2.0.b4
Appium-Python-Client>=2.6.1
python-dotenv==0.20.0